backend/requirements.txt
# backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.2.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
bcrypt==4.1.2
httpx==0.26.0
python-dotenv==1.0.1

# PostgreSQL support
asyncpg==0.29.0
psycopg2-binary==2.9.9

# LLM providers
openai==1.12.0
anthropic==0.18.0

# Document generation
python-docx==1.2.0
lxml==5.2.2
